Moved ViewToggle to the main folder
